refactor: 💡 Migrate DatePicker to MUI5 (#747)
* refactor: 💡 Migrate DatePicker to MUI5

✅ Closes: #745
