+++ notes/SQFormDatePicker.md
@@ -4,15 +4,15 @@ This is a date time field used inside our forms.
 
 ## First time setup
 
-`> npm install @material-ui/pickers@next`
+`> npm install @mui/x-date-pickers@next`
 
 From the top of your client code add the LocalizationProvider. The same place the Material-UI theme providers are used. Then pass the MomentAdapter into the `dateAdapter` prop.
 
 ```js
-import {LocalizationProvider} from '@material-ui/pickers';
-import MomentAdapter from '@material-ui/pickers/adapter/moment';
+import {LocalizationProvider} from '@mui/x-date-pickers';
+import {AdapterMoment} from '@mui/x-date-pickers/AdapterMoment';
 
-<LocalizationProvider dateAdapter={MomentAdapter} locale={'en'}>
+<LocalizationProvider dateAdapter={AdapterMoment} locale={'en'}>
   <App />
 </LocalizationProvider>;
 ```
