+++ stories/SQFormDatePickerWithDateFNS.stories.tsx
@@ -2,11 +2,11 @@ import * as Yup from 'yup';
 import React from 'react';
 import * as markdown from '../notes/SQFormDatePicker.md';
 import {SQFormDatePickerWithDateFNS as SQFormDatePickerComponent} from '../src';
-import {createDocsPage} from './utils/createDocsPage';
-import {SQFormStoryWrapper} from './components/SQFormStoryWrapper';
+import {createDocsPage} from '../old_stories/utils/createDocsPage';
+import {SQFormStoryWrapper} from '../old_stories/components/SQFormStoryWrapper';
 import type {Story, Meta} from '@storybook/react';
 import type {SQFormDatePickerDateFNSProps} from 'components/SQForm/SQFormDatePickerWithDateFNS';
-import type {SQFormStoryWrapperProps} from './components/SQFormStoryWrapper';
+import type {SQFormStoryWrapperProps} from '../old_stories/components/SQFormStoryWrapper';
 
 export type FormProps = {
   initialValues?: SQFormStoryWrapperProps['initialValues'];
@@ -81,9 +81,6 @@ DatePickerBefore2024.args = {
       .max(new Date('2024-01-01'), 'Date must be before 2024')
       .typeError('Invalid date'),
   }),
-  muiFieldProps: {
-    allowSameDateSelection: true,
-  },
 };
 
 export default meta;
